fix: update threshold data type on logpress config

+++ pkg/logpress/logpress_dto.go
@@ -5,7 +5,7 @@ import "time"
 type LogpressConfig struct {
 	Username          string    `json:"username"`
 	Password          string    `json:"password"`
-	Threshold         uint8     `json:"threshold"`
+	Threshold         uint16    `json:"threshold"`
 	ThresholdUnit     string    `json:"threshold_unit"`
 	LastDumpTimestamp time.Time `json:"last_dump_timestamp"`
 	CronInterval      uint8     `json:"cron_interval"`
